chore: split different commands into different classes so we don't have to instantiate things that don't need to be
