+++ Peer.Domain/Commands/ShowConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace Peer.Domain.Commands
+{
+    public class ShowConfig
+    {
+
+    }
+}
